Write some of the easy block_to_html functions
I've only tested the block_to_html_paragraph function so far, but I
believe that the others should be more or less working as expected.
I'll write out the tests for them in the morning. The functions for
creating the ordered and unordered lists will be a bit more involved.